refactor(usecases): add the users' use cases in a different folder
Refs #126
