+++ backend/src/main/config/dependencyInjection/userDependencyInjection.ts
@@ -1,16 +1,16 @@
 import { BcryptPassword } from "@/application/utils/hashUtils";
 import { CreateUserController } from "@/main/controllers/user/createUserController";
-import { CreateUserUseCase } from "@/application/useCases/createUseCases";
+import { CreateUserUseCase } from "@/application/useCases/user/createUseCases";
 import { DefaultCategoryService } from "@/application/services/defaultCategoryService";
 import { DeleteUserController } from "@/main/controllers/user/deleteUserController";
-import { DeleteUserUseCase } from "@/application/useCases/deleteUserUseCase";
+import { DeleteUserUseCase } from "@/application/useCases/user/deleteUserUseCase";
 import { JWTTokenGenerator } from "@/application/utils/authUtils";
 import { LoginUserController } from "@/main/controllers/user/loginUserController";
-import { LoginUserUseCase } from "@/application/useCases/loginUserUseCase";
+import { LoginUserUseCase } from "@/application/useCases/user/loginUserUseCase";
 import { PrismaAuthUser } from "@/infrastructure/database/prisma/prismaAuthUser";
 import { PrismaCategoryRepository } from "@/infrastructure/database/prisma/prismaCategoryRepository";
 import { UpdateUserController } from "@/main/controllers/user/updateUserController";
-import { UpdateUserUseCase } from "@/application/useCases/updateUserUseCase";
+import { UpdateUserUseCase } from "@/application/useCases/user/updateUserUseCase";
 import { UserFactory } from "@/domain/factories/userFactory";
 
 // Auth
